refactor: Move validator to more common base
